Removed .tar option
Sonatype Nexus IQ does not properly unarchive from a tar file, as it assumes that the tar file isa Docker imagee.  Changed this to archive to a .zip file instead.
The .xz is really optional, as it is only used to allow uploading to a customer site where .zip files are rejected.